chore: extract to models transaction result
