+++ backend/src/services/transactionManager/index.ts
@@ -1,4 +1,4 @@
+export { Transaction, TransactionResult } from "../../models/index.js";
 export { createKeyPair } from "../../utils/networkApi.js";
 export { getAccountNonce } from "./queue.js";
 export { createTransactionManager } from "./transactionManager.js";
-export { Transaction, TransactionResult } from "./types.js";
